Parameter checking: simply return on error, as is the idiom
if zend_parse_parameters() fails, it will print an error message saying
what kind of error occurred (maybe not enough parameters, maybe
parameters of the wrong type), and populate the return value with
`false`. What all the other extensions do in this situation, and what is
documented in the Oreilly PHP book, is to just return.

The current behavior is certainly wrong, because it will print a message
about wrong parameter counts regardless of what the real error is.