config.m4: rewrite, search for library and headers
The extension no longer compiled out of the box with Samba-4.x,
because Samba moved its include files from /usr/include to
/usr/include/samba-4.x. Add this new location to the list of search
paths.

Do the shared library lookup in the standard PHP way (gleaned from
several other extensions) to avoid using hardcoded paths and filenames
for the shared library (not portable).

Try to do something sensible with the DIR argument to
--with-libsmbclient=DIR: if not given, the system libraries are used,
else the script expects the library to be in DIR/lib, and the headers in
DIR/include.